Перенос конфигурации DbConnection в extention метод
